feat(brave): delegated DHT + preload
- enabled delegated DHT
- disabled -star as it crash boot if server is down
- enabled preload (best-effort)
- enabled mdns
- simplified configuration
