+++ add-on/_locales/en/messages.json
@@ -236,7 +236,7 @@
     "description": "An option title on the Preferences screen (option_ipfsNodeConfig_title)"
   },
   "option_ipfsNodeConfig_description": {
-    "message": "Additional configuration for the embedded IPFS node (arrays will be merged). Must be valid JSON.",
+    "message": "Additional configuration for the embedded JS IPFS node. Must be valid JSON.",
     "description": "An option description on the Preferences screen (option_ipfsNodeConfig_description)"
   },
   "option_ipfsNodeType_external": {
