+++ package.json
@@ -116,7 +116,7 @@
     "@material/switch": "3.1.0",
     "browser-process-hrtime": "1.0.0",
     "choo": "7.0.0",
-    "chrome-dgram": "https://github.com/lidel/chrome-dgram/tarball/2bd217a29ea91b313fe5f46828621bb27d69cb06/chrome-dgram.tar.gz",
+    "chrome-dgram": "3.0.3",
     "chrome-net": "3.3.2",
     "debug": "4.1.1",
     "doc-sniff": "1.0.1",
@@ -126,7 +126,7 @@
     "get-port": "5.0.0",
     "http-dns": "3.0.1",
     "http-node": "1.2.0",
-    "ipfs": "https://github.com/ipfs/js-ipfs/tarball/ddd49ce41f6f9abd48b1e54bc4c7dd3e58fd0ee9/js-ipfs.tar.gz",
+    "ipfs": "https://github.com/ipfs/js-ipfs/tarball/cf38aead2b0cb0b5f269daf265a2b868c50a81f8/js-ipfs.tar.gz",
     "ipfs-css": "0.13.1",
     "ipfs-http-client": "33.1.1",
     "ipfs-http-response": "0.3.1",
